==> Introduction to Data Analysis/3. Case study 2/1.assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd

df_08 = pd.read_csv('all_alpha_08.csv')
df_18 = pd.read_csv('all_alpha_18.csv')

df_08.head()

df_18.head()

**Quiz questions**
* number of samples in each dataset
* number of columns in each dataset
* duplicate rows in each dataset
* datatypes of columns
* features with missing values
* number of non-null unique values for features in each dataset
* what those unique values are and counts for each

# number of samples & columns in df_08
df_08.shape

# number of duplicated rows in df_08
sum(df_08.duplicated())

# rows with missing data in df_08
df_08.isnull().sum()

# number of samples & columns in df_18
df_18.shape

# number of duplicated rows in df_18
sum(df_18.duplicated())

# rows with missing data in df_18
df_18.isnull().sum()

# data types of columns in df_08
df_08.dtypes

# data types of columns in df_18
df_18.dtypes

df_08.SmartWay.nunique(dropna=True)

df_18.SmartWay.nunique(dropna=True)

df_08["Sales Area"].nunique()

df_18["Cert Region"].nunique()

df_08.Trans.nunique()

df_08['Cyl'].nunique()

df_18['Cyl'].nunique()

df_08.Fuel.unique()

df_18.Fuel.unique()
